+++ src/base/light_sampler.rs
@@ -2,9 +2,9 @@ use crate::pbrt::*;
 
 pub struct SampledLight {
     pub light: Arc<dyn Light>,
-    pub p: Float,
+    pub p: f64,
 }
 
 pub trait LightSampler {
-    fn sample(&self, u: Float) -> Option<SampledLight>;
+    fn sample(&self, u: f64) -> Option<SampledLight>;
 }
